+++ src/evm_transition_tool/__init__.py
@@ -7,6 +7,14 @@ from .evmone import EvmOneTransitionTool
 from .execution_specs import ExecutionSpecsTransitionTool
 from .geth import GethTransitionTool
 from .nimbus import NimbusTransitionTool
+from .traces import (
+    EVMCallFrameEnter,
+    EVMCallFrameExit,
+    EVMTraceLine,
+    EVMTransactionTrace,
+    TraceableException,
+    TraceMarkerDescriptor,
+)
 from .transition_tool import (
     FixtureFormats,
     TransitionTool,
@@ -18,11 +26,17 @@ TransitionTool.set_default_tool(GethTransitionTool)
 
 __all__ = (
     "BesuTransitionTool",
+    "EVMCallFrameEnter",
+    "EVMCallFrameExit",
     "EvmOneTransitionTool",
+    "EVMTraceLine",
+    "EVMTransactionTrace",
+    "TraceMarkerDescriptor",
     "ExecutionSpecsTransitionTool",
     "FixtureFormats",
     "GethTransitionTool",
     "NimbusTransitionTool",
+    "TraceableException",
     "TransitionTool",
     "TransitionToolNotFoundInPath",
     "UnknownTransitionTool",